+++ src/Telegram.Bot/Requests/Available Methods/Admin Bot/KickChatMemberRequest.cs
@@ -23,7 +23,7 @@ namespace Telegram.Bot.Requests
         /// Unique identifier of the target user
         /// </summary>
         [JsonProperty(Required = Required.Always)]
-        public int UserId { get; }
+        public long UserId { get; }
 
         /// <summary>
         /// Date when the user will be unbanned, unix time. If user is banned for more than 366 days or less than 30 seconds from the current time they are considered to be banned forever.
@@ -43,7 +43,7 @@ namespace Telegram.Bot.Requests
         /// </summary>
         /// <param name="chatId">Unique identifier for the target group or username of the target supergroup or channel</param>
         /// <param name="userId">Unique identifier of the target user</param>
-        public KickChatMemberRequest(ChatId chatId, int userId)
+        public KickChatMemberRequest(ChatId chatId, long userId)
             : base("kickChatMember")
         {
             ChatId = chatId;
